backend/requirements.txt
django==6.0.4
djangorestframework==3.15.2
psycopg2-binary==2.9.12
celery==5.6.3
redis==7.4.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-asgiref==3.11.1
django-sqlparse==0.5.5
django-tzlocal==5.3.1
billiard==4.2.4
vine==5.1.0
amqp==5.3.1
kombu==5.6.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
